pyproject.toml
[project]
name = "so-arm101"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pyserial>=3.5",
]
